Add helper to get table creation query
