+++ workers/loc.api/sync/dao/helpers/index.js
@@ -21,6 +21,7 @@ const filterModelNameMap = require('./filter-model-name-map')
 const SQL_OPERATORS = require('./sql.operators')
 const getSymbolFilter = require('./get-symbol-filter')
 const getStatusMessagesFilter = require('./get-status-messages-filter')
+const getTableCreationQuery = require('./get-table-creation-query')
 
 module.exports = {
   mixUserIdToArrData,
@@ -39,5 +40,6 @@ module.exports = {
   filterModelNameMap,
   SQL_OPERATORS,
   getSymbolFilter,
-  getStatusMessagesFilter
+  getStatusMessagesFilter,
+  getTableCreationQuery
 }
